Add attribute to limit the canvas-height to its container
This patch adds an attribute that, combined with `scale='page-fit'`,
allows to fit the contents inside of the page both with respects to the
width and with respect to the height. Think of it as using
`background-size='contain'` rather than the default behavior of
`background-size='cover'` that `scale='page-fit'` triggers.